refactor(auth): Improve structure and enhance security with serialization
- Added interceptors to streamline request and response handling
- Implemented serialization to exclude passwords in responses
- Separated login and register logic into dedicated auth provider methods
- Enhanced code readability and maintainability
